Simplify the lexer
I had an idea of a state machine where each state was a class and the lexer would bounce between them, but there are really only two non-trivial states besides the main state, so that wasn't a very useful abstraction.

This version simply gobbles up the next token from the character stream when asked to.