Add a native compiler
Since the #to_s representation of the parse tree is valid Ruby syntax we can stick it inside of a Proc and use named arguments to pass in the bindings for the references.

This will only work in Rubies that support named arguments.

I experimented with Binding and #local_variable_set, but didn't get it to work unfortunately.